util: refactor getClientIP to extract only IP addr from RemoteAddr
Previously the function returned the entire `RemoteAddr` field which includes
both IP address and port number, but now we are using `strings.Split` function
to split the `RemoteAddr` field by `:` and get the first value which is the IP
address. This change will make the function more accurate and reliable.