chore: handle file not found error when getResources
This commit message documents changes made to teler.go
file that handles the file not found error when reading
the threat resources. Previously, the program returned
an error immediately when it encountered a file not
found error. With this change, the program will check
if the error is a file not found error and attempt to
retrieve the threat before retrying to read the file.
If the threat retrieval fails, an error will be returned.
Otherwise, the program will continue to read the file
as usual.

These changes improve the robustness of the program by
gracefully handling file not found errors when reading
the threat resources.