feat: add in-memory opt to load threat dataset on initialization
This commit adds a new boolean option, `InMemory`, to the `teler.Teler` middleware.
When `InMemory` is set to true, the threat dataset will be loaded into memory on
initialization, which can improve performance when running the middleware on a
distroless or runtime image. Note that this option may not be suitable for larger
datasets, as it may consume a significant amount of memory. When `InMemory` is set
to false, the threat dataset will be downloaded and stored under the user-level
cache directory on the first request, and subsequent requests will use the cached
dataset. Close #23.